fix m_keterangan_rapor migration

+++ database/migrations/1621036028314_m_keterangan_rapor_schema.js
@@ -5,18 +5,21 @@ const Schema = use("Schema");
 
 class MKeteranganRaporSchema extends Schema {
   up() {
-    this.create("m_keterangan_rapor", (table) => {
-      table.increments();
-      table.integer("m_user_id").unsigned().index("m_user_id");
-      table
-        .foreign("m_user_id")
-        .references("m_user.id")
-        .onDelete("cascade")
-        .onUpdate("cascade");
-      table.string("catatan");
-      table.string("kelulusan");
-      table.boolean("dihapus").defaultTo(0);
-      table.timestamps();
+    this.hasTable("m_keterangan_rapor").then((exist) => {
+      if (!exist)
+        this.create("m_keterangan_rapor", (table) => {
+          table.increments();
+          table.integer("m_user_id").unsigned().index("m_user_id");
+          table
+            .foreign("m_user_id")
+            .references("m_user.id")
+            .onDelete("cascade")
+            .onUpdate("cascade");
+          table.string("catatan");
+          table.string("kelulusan");
+          table.boolean("dihapus").defaultTo(0);
+          table.timestamps();
+        });
     });
   }
